requirement.txt
langchain 
wikipedia 
langchain-community 
langchain-openai
langchain-anthropic
langchain-mistral
python-dotenv
pydantic
duckduckgo-search
